+++ src/main/java/com/ormi/cookudasse/post/entitiy/FoodCategory.java
@@ -1,4 +1,4 @@
-package com.ormi.cookudasse.entitiy;
+package com.ormi.cookudasse.post.entitiy;
 
 public enum FoodCategory {
     SOUP, RICE, DESSERT, NOODLE
